% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt1_baixar_sequencia.R
\name{trt1_baixar_sequencia}
\alias{trt1_baixar_sequencia}
\title{Baixa julgados de primeiro e segundo grau do TRT1 pela sequência}
\usage{
trt1_baixar_sequencia(x, diretorio = ".")
}
\arguments{
\item{x}{Número da sequência}

\item{diretorio}{Diretório}
}
\value{
html
}
\description{
Baixa julgados de primeiro e segundo grau do TRT1 pela sequência
}
