% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt1_ler_sequencia.R
\name{trt1_ler_sequencia}
\alias{trt1_ler_sequencia}
\title{Lê metadados das decisões do TRT1 baixados com trt1_baixar_sequencia}
\usage{
trt1_ler_sequencia(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar diretório se não informar arquivos}
}
\value{
tibble
}
\description{
Lê metadados das decisões do TRT1 baixados com trt1_baixar_sequencia
}
